Refactor server tests.
