+++ package.json
@@ -42,6 +42,7 @@
     "istanbul": "^0.4.3",
     "snazzy": "^7.0.0",
     "standard": "^10.0.2",
+    "strip-ansi": "^3.0.1",
     "tape": "^4.6.3"
   }
 }
