+++ src/lib/world.js
@@ -4,8 +4,7 @@ const WorldObjectClassBuilder = require('./world-object-class-builder')
 const WorldObjectProxyBuilder = require('./world-object-proxy-builder')
 const idify = require('./idify')
 const Deserializer = require('./deserializer')
-const parse = require('./parse').parseSentence
-const noun = require('./parse').parseNoun
+const { parseSentence: parse, parseNoun: noun } = require('./parse')
 const { color } = require('./colors')
 const NamespaceNode = require('./namespace-node')
 const Timer = require('./timer')
