+++ test/controllers/socket-controller-test.js
@@ -1,10 +1,6 @@
 const test = require('tape')
+const stripAnsi = require('strip-ansi')
 const SocketController = require('../../src/controllers/socket-controller')
-
-// From 'chalk':
-// use bright blue on Windows as the normal blue color is illegible
-// i.e. escape 94m instead of 34m.
-var isSimpleWindowsTerm = process.platform === 'win32' && !/^xterm/i.test(process.env.TERM)
 
 function loggerFactory () {
   return {
@@ -18,7 +14,7 @@ function socketFactory (id = '1234') {
     events: [],
     id: id,
     on () {},
-    emit (event, msg) { this.events.push([event, msg]) },
+    emit (event, msg) { this.events.push([event, stripAnsi(msg)]) },
     lastEvent () { return this.events[this.events.length - 1] }
   }
 }
@@ -38,9 +34,7 @@ test('SocketController: emits a welcome message when a socket connects', t => {
 
   const expected = [
     'output',
-    isSimpleWindowsTerm
-      ? 'Welcome to \x1b[1m\x1b[94mroom.js\x1b[39m\x1b[22m!\nType \x1b[1m\x1b[35mhelp\x1b[39m\x1b[22m for a list of available commands.'
-      : 'Welcome to \x1b[1m\x1b[34mroom.js\x1b[39m\x1b[22m!\nType \x1b[1m\x1b[35mhelp\x1b[39m\x1b[22m for a list of available commands.'
+    'Welcome to room.js!\nType help for a list of available commands.'
   ]
 
   t.deepEqual(socket.lastEvent(), expected)
